psu version 0.2.0-alpha.15
'--compose-file' option is back and '--stack-file' option is an alias

Rename 'system' action to 'system:info'

Add '--masked-variables' option: in debug and/or verbose mode, the value of sensitive variables will be hidden, useful to avoid leaking passwords or tokens in your logs. Possible values: true|extended|false. Defaults to 'true'.
Useful when running 'psu' in a CI tool, to avoid leaking passwords or tokens in the logs.